requirements.txt
# These requirements have been set as of 03/02/2026, newer versions could certainly exist.
# for data handling
pandas>=2.2.0
numpy>=1.26.4
pyarrow>=14.0.2
matplotlib>=3.8.2
seaborn>=0.13.1
scikit-learn>=1.3.2
imbalanced-learn
xgboost>=2.0.3
lightgbm
fastapi>=0.109.0
uvicorn>=0.27.0
mlflow>=2.9.2
loguru>=0.7.0 # Could also use 'logging', if preferred
boto3>=1.34.34 # I am using AWS, this won't be needed if you deploy to GCP or Azure..
python-dotenv
fsspec
s3fs
setuptools
pytest>=7.4.4
pytest-loguru>=0.4.0
pyyaml>=6.0.1